Improve performance of tests related to docs gen
